utils/market_intel.py
import requests

def get_live_job_trends(role):
    """
    Mock integration for an API like SerpApi or Adzuna.
    In production, this would fetch real skills trending for the role.
    """
    # Example: Fetching trending skills for a 'Backend Developer'
    api_url = f"https://api.example.com/trends?role={role}"
    # response = requests.get(api_url)
    # return response.json()['trending_skills']
    return ["Docker", "Kubernetes", "GraphQL"] # Mock data for demo