requirements.txt
streamlit
groq
python-dotenv
chromadb
sentence-transformers
pdfplumber